Delete BaseAdvice and rename BaseInterceptor to the AbstractInterceptor
