+++ tests/Go/Stubs/AbstractInterceptorMock.php
